requirements.txt
asgiref==3.7.2
Django==3.0.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
django-cors-headers==3.2.1
#psycopg2-binary==2.8.6

